refactor(runtime): isolate session stop handling
