+++ apps/desktop/src-tauri/crates/host-application/src/app_service/build_orchestrator/mod.rs
@@ -1,6 +1,7 @@
 mod build_cleanup;
 mod build_lifecycle;
 mod build_runtime_setup;
+mod session_stop;
 
 use serde::Deserialize;
